+++ CHANGELOG.md
@@ -51,6 +51,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
   * `getTimestampHex()`
   * `getVariant()`
   * `getVersion()`
+* Remove the following deprecated constants:
+  * `Ramsey\Uuid\Uuid::VALID_PATTERN`
 * Remove dependency on ramsey/collection package.